调整数据模型：将 categoryId 类型改为 Integer，更新图片字段处理逻辑，适配前端需求

+++ src/main/java/com/campus/trade/entity/User.java
@@ -5,14 +5,14 @@ import java.time.LocalDateTime;
 
 @Data
 public class User {
-    private Integer id;           // 数据库是 INT
-    private String studentId;     // 学号
+    private Long id; // 改为 Long，与数据库 BIGINT 一致
+    private String studentId; // 学号
     private String username;
-    private String password;      // 加密密文
+    private String password; // 加密密文
     private String phone;
-    private String avatar;        // 头像 URL
-    private Integer status;       // 状态：0-禁用，1-正常
-    private Integer isAuth;       // 是否实名认证：0-否，1-是 (新增)
+    private String avatar; // 头像 URL
+    private Integer status; // 状态：0-禁用，1-正常
+    private Integer isAuth; // 是否实名认证：0-否，1-是 (新增)
     private LocalDateTime createTime;
     private LocalDateTime updateTime;
     private Integer creditScore;
